effects: move GunShot_Spawn
