+++ src/game/objects/ai/cowboy.c
@@ -2,9 +2,9 @@
 
 #include "game/box.h"
 #include "game/collide.h"
+#include "game/effects/gunshot.h"
 #include "game/items.h"
 #include "game/lot.h"
-#include "game/objects/effects/gunshot.h"
 #include "game/people.h"
 #include "global/vars.h"
 
@@ -133,7 +133,7 @@ void CowboyControl(int16_t item_num)
                     }
                 } else {
                     int16_t fx_num =
-                        CreatureEffect(item, &g_CowboyGun2, GunShot_Spawn);
+                        CreatureEffect(item, &g_CowboyGun2, Effect_GunShot);
                     if (fx_num != NO_ITEM) {
                         g_Effects[fx_num].pos.y_rot += head;
                     }
